classe ReajusteService se tornou muito mais extensível com a implementação de uma abstração(interface) para gerenciar as validacoes
