+++ src/br/com/alura/rh/service/ValidacaoReajuste.java
@@ -1,0 +1,10 @@
+package br.com.alura.rh.service;
+
+import br.com.alura.rh.model.Funcionario;
+
+import java.math.BigDecimal;
+
+public interface ValidacaoReajuste {
+
+    void validar(Funcionario funcionario, BigDecimal aumento);
+}
